refactor: moved auth variables
